CMake build for dynamic linking with system libraries
Compiles main.cpp and dependencies of SDL2 and Boost's program_options
as long as the packages are installed on the user's Linux pc.

Untested on a Windows build -- most likely not supported currently.

Will get a working build for distribution where linking executable with
the shared library will be relative for the project hierarchy within
the next couple of commits.
Will add test target for running tests and doc target for generating
documentation.